2 Var Bank Grapher 1.2
So I didn't call the last one 1.1, but at this point, the program now saves the data to a text file, which can be inputted into a excel file. I would go directly to the .csv file, but when storing the fraction of: (people who paid off loans)/(all loaners in this income and price range), excel would store the 1/1 as January 1st. It's fairly legible in it's current state, but over the next 45 minutes, I'm going to work on adding units to the vertical and horizontal axes.