2 Variable Bank Grapher 1.3
This is the final version of the 2 Variable Bank Grapher being made at HackUTD '19. Right now, the data is sent into a .txt format, where it can be read into an excel file for easy viewing. There are also axes that let you know the value of either changed quantity in the .txt file. If none of this makes sense, then I truly hope I can clear things up in my demo.